java移植作業
git-svn-id: https://svn.sourceforge.jp/svnroot/cadencii/Cadencii/trunk@698 9d680d4f-cbfa-477e-9905-73dbafb9b024

+++ Cadencii/OtoArgs.cs
@@ -25,11 +25,11 @@ namespace Boare.Cadencii {
     public struct OtoArgs {
         public String fileName;
         public String Alias;
-        public int msOffset;
-        public int msConsonant;
-        public int msBlank;
-        public int msPreUtterance;
-        public int msOverlap;
+        public float msOffset;
+        public float msConsonant;
+        public float msBlank;
+        public float msPreUtterance;
+        public float msOverlap;
     }
 
 #if !JAVA
